requirements.txt
Django
django-environ
requests
Pillow
psutil
psycopg2-binary